Small DB model updates and world seed data

+++ app/TW.Vault.Lib/Features/Simulation/LoyaltyCalculator.cs
@@ -7,16 +7,16 @@ namespace TW.Vault.Features.Simulation
 {
     public class LoyaltyCalculator
     {
-        private int loyaltyPerHour;
+        private decimal loyaltyPerHour;
 
-        public LoyaltyCalculator(int loyaltyPerHour)
+        public LoyaltyCalculator(decimal loyaltyPerHour)
         {
             this.loyaltyPerHour = loyaltyPerHour;
         }
 
         public short PossibleLoyalty(short oldLoyalty, TimeSpan timeSinceLoyalty)
         {
-            var result = (short)Math.Min(100, Math.Abs(oldLoyalty) + timeSinceLoyalty.TotalHours * loyaltyPerHour);
+            var result = (short)Math.Min(100, Math.Abs(oldLoyalty) + timeSinceLoyalty.TotalHours * (double)loyaltyPerHour);
 
             return result;
         }
